requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
graphviz==0.20.1
lxml==5.1.0
pyyaml==6.0.1
click==8.1.7
pytest==8.0.0 